# src/Wolf.Notification.EmailSender/OpenAPIs/NotifApiSwagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Wolf Notification API",
    "description": "Wolf API for Managing Notification Templates and Messages and for Message-delivery",
    "contact": {
      "name": "API",
      "url": "https://www.zflows.com/"
    },
    "license": { },
    "version": "v2.0"
  },
  "paths": {
...
    "/template": {
      "get": {
        "tags": [
          "Template"
        ],
        "summary": "Get All Templates",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.TemplateWIdDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.TemplateWIdDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.TemplateWIdDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "OAuth2": [
              "OAuth2"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "Template"
        ],
        "summary": "Create new Template",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.TemplateDto"
                  }
                ],
                "description": "",
                "nullable": true
              }
            },
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.TemplateDto"
                  }
                ],
                "description": "",
                "nullable": true
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.TemplateDto"
                  }
                ],
                "description": "",
                "nullable": true
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.TemplateDto"
                  }
                ],
                "description": "",
                "nullable": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.TemplateWIdDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.TemplateWIdDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.TemplateWIdDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "OAuth2": [
              "OAuth2"
            ]
          }
        ]
      }
    },
    "/template/{id}": {
      "get": {
        "tags": [
          "Template"
        ],
        "summary": "Get Template by Id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "description": "",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.TemplateWIdDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.TemplateWIdDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.TemplateWIdDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "OAuth2": [
              "OAuth2"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "Template"
        ],
        "summary": "Update Template",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.TemplateDto"
                  }
                ],
                "nullable": true
              }
            },
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.TemplateDto"
                  }
                ],
                "nullable": true
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.TemplateDto"
                  }
                ],
                "nullable": true
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.TemplateDto"
                  }
                ],
                "nullable": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.TemplateWIdDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.TemplateWIdDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.TemplateWIdDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "OAuth2": [
              "OAuth2"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Template"
        ],
        "summary": "Delete Template by Id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "description": "",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "OAuth2": [
              "OAuth2"
            ]
          }
        ]
      }
    },
    "/template/{id}/recipient": {
      "post": {
        "tags": [
          "Template"
        ],
        "summary": "Add a recipient to Template",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "description": "",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.RecipientWTypeDto"
                  }
                ],
                "description": "",
                "nullable": true
              }
            },
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.RecipientWTypeDto"
                  }
                ],
                "description": "",
                "nullable": true
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.RecipientWTypeDto"
                  }
                ],
                "description": "",
                "nullable": true
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.RecipientWTypeDto"
                  }
                ],
                "description": "",
                "nullable": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.RecipientWIdDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.RecipientWIdDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.RecipientWIdDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "OAuth2": [
              "OAuth2"
            ]
          }
        ]
      }
    },
    "/template/{id}/recipient/{recipientId}/{typeCode}": {
      "delete": {
        "tags": [
          "Template"
        ],
        "summary": "Delete a Recipient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "description": "",
              "format": "uuid"
            }
          },
          {
            "name": "recipientId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int64"
            }
          },
          {
            "name": "typeCode",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "description": "",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "OAuth2": [
              "OAuth2"
            ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "Wolf.Notification.Models.RecipientDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Wolf.Notification.Models.RecipientWTypeDto": {
        "type": "object",
        "properties": {
          "typeCode": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Wolf.Notification.Models.RecipientWIdDto": {
        "type": "object",
        "properties": {
          "recipientId": {
            "type": "integer",
            "format": "int64"
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Wolf.Notification.Models.RecipientWIdTypeDto": {
        "type": "object",
        "properties": {
          "typeCode": {
            "type": "string",
            "nullable": true
          },
          "recipientId": {
            "type": "integer",
            "format": "int64"
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Wolf.Notification.Models.TemplateWIdDto": {
        "type": "object",
        "properties": {
          "templateId": {
            "type": "string",
            "format": "uuid"
          },
          "defaultFromRecipient": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Wolf.Notification.Models.RecipientWIdDto"
              }
            ],
            "nullable": true
          },
          "defaultRecipients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Wolf.Notification.Models.RecipientWIdTypeDto"
            },
            "nullable": true
          },
          "templateName": {
            "type": "string",
            "nullable": true
          },
          "templateSubject": {
            "type": "string",
            "nullable": true
          },
          "templateBody": {
            "type": "string",
            "nullable": true
          },
          "defaultProviderCode": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Wolf.Notification.Models.TemplateDto": {
        "type": "object",
        "properties": {
          "defaultFromRecipient": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Wolf.Notification.Models.RecipientDto"
              }
            ],
            "nullable": true
          },
          "defaultRecipients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Wolf.Notification.Models.RecipientWTypeDto"
            },
            "nullable": true
          },
          "templateName": {
            "type": "string",
            "nullable": true
          },
          "templateSubject": {
            "type": "string",
            "nullable": true
          },
          "templateBody": {
            "type": "string",
            "nullable": true
          },
          "defaultProviderCode": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "OAuth2": {
        "type": "oauth2",
        "description": "Balea Server OpenId Security Scheme",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://devid.mycompany.com/connect/authorize",
            "tokenUrl": "https://devid.mycompany.com/connect/token",
            "scopes": {
              "notif_api": "Wolf Notification Api Audience",
              "openid": "OpenID"
            }
          }
        }
      }
    }
  }
}